Break out club application form into component
